Refactoring: renamed component from OrderBookWidget to OrderBookChart
