+++ src/components/OrderBookChart/chartFunctions.ts
