+++ src/components/OrderBookChart/index.tsx
@@ -18,10 +18,10 @@ import { TEN_BIG_NUMBER, DEFAULT_PRECISION, ZERO_BIG_NUMBER, ORDERBOOK_DATA_FETC
 import BN from 'bn.js'
 import { useDebounce } from 'hooks/useDebounce'
 
-import { Offer, PricePointDetails, ZoomValues } from 'components/OrderBookWidget/types'
-import { calcInitialZoom, calcZoomY } from 'components/OrderBookWidget/zoomFunctions'
-import { createChart, getZoomButtonContainer, setLabel } from 'components/OrderBookWidget/chartFunctions'
-import { processData, _printOrderBook } from 'components/OrderBookWidget/dataProcessingFunctions'
+import { Offer, PricePointDetails, ZoomValues } from 'components/OrderBookChart/types'
+import { calcInitialZoom, calcZoomY } from 'components/OrderBookChart/zoomFunctions'
+import { createChart, getZoomButtonContainer, setLabel } from 'components/OrderBookChart/chartFunctions'
+import { processData, _printOrderBook } from 'components/OrderBookChart/dataProcessingFunctions'
 
 const ZOOM_INCREMENT_PERCENTAGE = 0.25 // %
